Paginate orders screen and implement product category screen

+++ app/src/main/java/com/shahid/fashionista_mobile/dto/response/ProductListResponse.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductListResponse {
-    private long total;
-    private long current;
+    private int total;
+    private int current;
     private List<ProductResponse> products = new ArrayList<>();
 }
